+++ src/features/diet-study-playback/components/diet-score-header/index.tsx
