+++ src/features/diet-study-playback/components/gut-score/score-range.tsx
