+++ src/features/diet-study-playback/components/quality-score/index.tsx
