+++ src/features/diet-study-playback/components/quality-score/styles.ts
